requirements.txt
langchain
langgraph
langchain-community
chromadb
pypdf
ollama
python-dotenv
fastapi
uvicorn
tqdm
